-updated the domain entities

+++ Fulfill_RepositoryService/DomainEntities/Priority.cs
@@ -8,6 +8,7 @@ namespace Fulfill_RepositoryService.DomainEntities
     {
  
         public Guid PriorityId { get; set; }
-        public int Merit { get; set; }
+        public decimal Merit { get; set; }
+
     }
 }
